requirements.txt
celery==3.1.18
tornado==4.1
tornado-redis==2.4.18
redis==2.10.3
mapnik2==2.2.0
Pillow==2.8.1
protobuf==2.6.1